un-breaking decimal transformation

+++ src/parser/transform/helpers/transform_typename.cpp
@@ -186,7 +186,7 @@ LogicalType Transformer::TransformTypeName(duckdb_libpgquery::PGTypeName &type_n
 			if (scale > width) {
 				throw ParserException("Scale cannot be bigger than width");
 			}
-			result_type = LogicalType::DECIMAL(NumericCast<int>(width), NumericCast<int>(scale));
+			result_type = LogicalType::DECIMAL(NumericCast<uint8_t>(width), NumericCast<uint8_t>(scale));
 			break;
 		case LogicalTypeId::INTERVAL:
 			if (modifier_idx > 1) {
